Sample into data set (subscription notification) instead of single value/subscription notifications
