+++ src/Azure.IIoT.OpcUa.Publisher/tests/Services/Encoder/NetworkMessage.cs
@@ -142,17 +142,17 @@ namespace Azure.IIoT.OpcUa.Publisher.Tests.Services
             var seq = 1u;
 
 #pragma warning disable CA2000 // Dispose objects before losing scope
-            var dataItem = new OpcUaMonitoredItem.DataItem(new DataMonitoredItemModel
+            var dataItem = new OpcUaMonitoredItem.DataChange(new DataMonitoredItemModel
             {
                 StartNodeId = "i=2258"
-            }, Log.Console<OpcUaMonitoredItem.DataItem>());
+            }, Log.Console<OpcUaMonitoredItem.DataChange>());
 #pragma warning restore CA2000 // Dispose objects before losing scope
 #pragma warning disable CA2000 // Dispose objects before losing scope
-            var eventItem = new OpcUaMonitoredItem.EventItem(new EventMonitoredItemModel
+            var eventItem = new OpcUaMonitoredItem.Event(new EventMonitoredItemModel
             {
                 StartNodeId = "i=2258",
                 EventFilter = new EventFilterModel()
-            }, Log.Console<OpcUaMonitoredItem.EventItem>());
+            }, Log.Console<OpcUaMonitoredItem.Event>());
 #pragma warning restore CA2000 // Dispose objects before losing scope
             eventItem.Fields.Add(("1", default));
             eventItem.Fields.Add(("2", default));
